new restructuring
I went back to keeping the client and server in the same project.
I decided to do this because of some details related to the hackclub.
Also, there are many packages that would be repeated between the two.

I improved and fixed the ping/pong logic. There was an error in this system that I hadn't noticed.